:recycle: :: refactor components call with easier alias
