+++ src/app/components/thumbnail/large/index.jsx
