+++ src/app/components/thumbnail/normal/index.jsx
